Tr2 completo
Adicionada BlockTileSet. Trabalho 2 finalizado

+++ src/TileSet.h
@@ -8,7 +8,7 @@ class TileSet
 public:
     virtual ~TileSet();
     virtual void open(const std::string&) = 0;
-    virtual void render(unsigned,int,int) = 0;
+    virtual void render(unsigned,float,float) = 0;
     int getTileWidth();
     int getTileHeight();
 protected:
